% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_function.R
\name{first_function}
\alias{first_function}
\title{Get summary data for a stock}
\usage{
first_function(stock, year)
}
\arguments{
\item{stock}{a stock name, e.g. cod-347d, or cod to find all cod stocks, or NULL (default)
to process all stocks.}

\item{year}{the assessment year, e.g. 2015, or 0 to process all years.}
}
\value{
A dataframe.
}
\description{
Find the summary data for a specific stock in a given assessment year.
}
\details{
This function retrieves summary data for a specified stock and year from the
SAG and SD Databases. It first validates the stock code against the ICES vocabulary
database, then fetches stock information from the Stock Data (SD) database,
and the summary table data from the Stock Assessment Graphs (SAG) database.

The results from both sources are combined into a single dataframe which is returned
to the user.
}
\examples{

assessment_summary <- first_function("had.27.46a20", 2023)
head(assessment_summary)

}
\seealso{
\code{\link[icesSAG]{getSAG}} gets summary table from SAG DB.
\code{\link[icesSAG]{icesSAG-package}} gives an overview of the icesSAG package.

\code{\link{icesUtils-package}} gives an overview of the icesUtils package.
}
\author{
Colin Millar.
}
