% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesUtils-package.R
\docType{package}
\name{icesUtils-package}
\alias{icesUtils-package}
\alias{icesUtils}
\title{Functions for processing ICES data and producing ICES shiny applications}
\description{
A collection of functions useful for processing ICES data and making
ICES shiny apps.
}
\details{
\emph{Group of functions:}
\tabular{ll}{
  \code{\link{first_function}}             \tab example function\cr
}
}
\references{
ICES Stock Assessment Graphs database: \url{http://sg.ices.dk}.

ICES Stock Assessment Graphs web services:
\url{http://sg.ices.dk/webservices.aspx}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ices-tools-dev/icesUtils}
  \item \url{https://ices-tools-prod.r-universe.dev/icesUtils}
  \item Report bugs at \url{https://github.com/ices-tools-dev/icesUtils/issues}
}

}
\author{
\strong{Maintainer}: Colin Millar \email{colin.millar@ices.dk}

Authors:
\itemize{
  \item Luca Lamoni
  \item Neil Maginnes
}

}
