% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{image_fullscreen_on_click_js}
\alias{image_fullscreen_on_click_js}
\title{JavaScript to view image fullscreen on click}
\usage{
image_fullscreen_on_click_js()
}
\value{
A script tag containing JavaScript code.
}
\description{
Adds a JavaScript handler that lets users click on an image to enter/exit 
fullscreen
}
\examples{
if (requireNamespace("shiny", quietly = TRUE) && interactive()) {
  library(shiny)
  
  addResourcePath(
    "mypkg",
    system.file("www", package = "icesUtils")
  )
  
  ui <- navbarPage(
  header = image_fullscreen_on_click_js(),
  title = "Example",
  "Click for full screen",
  tags$img(src = "mypkg/example-image.png", id = "demo-img", width = "300px", height = "300px")
  )

server <- function(input, output, session) {}

shinyApp(ui, server)
}
}
