% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{navbar_dropdown_autoclose_js}
\alias{navbar_dropdown_autoclose_js}
\title{JavaScript to close navbar dropdown after navigation via actionbuttons or other non-navbar input}
\usage{
navbar_dropdown_autoclose_js()
}
\value{
A script tag containing JavaScript code.
}
\description{
Adds a JavaScript handler that closes open navbar dropdown menus
when clicking outside the navbar.
}
\examples{
if (requireNamespace("shiny", quietly = TRUE) && interactive()) {
  library(shiny)
  ui <- 
navbarPage(header = navbar_dropdown_autoclose_js(),
          id = "main-navbar",
          title = "Example",
          navbarMenu(
            "Menu",
            tabPanel("A", value = "tab_a", actionButton("navigate2B", "Go to Tab B")),
            tabPanel("B", value = "tab_b", actionButton("navigate2A", "Go to Tab A"))
          )
)


server <- function(input, output, session) {
 observeEvent(input$navigate2B, {
   updateNavbarPage(session, inputId = "main-navbar", selected = "tab_b")
 })
 
 observeEvent(input$navigate2A, {
   updateNavbarPage(session, inputId = "main-navbar", selected = "tab_a")
 })
}

shinyApp(ui, server)
}
}
