% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_text.R
\name{prepare_nested_text_lists_from_excel}
\alias{prepare_nested_text_lists_from_excel}
\title{Create a named nested list of dataframes containing text content, from excel files and a vector of names.}
\usage{
prepare_nested_text_lists_from_excel(paths, names)
}
\arguments{
\item{paths}{valid paths provided either as vector or list}

\item{names}{list or vector of characters}
}
\description{
Uses \code{prepare_text_from_excel} to make a nested list containing text 
content extractable by \emph{section}. Requires that \emph{paths} and 
\emph{names} are vectors of equal length. 
Output works with \code{select_text} where \emph{names} gives the 
\emph{top_level} parameter to \code{select_text}.  vector of names an excel file into a list of dataframes. Works with \code{usethis::use_data} and
 \code{select_text} to provide a simple means of incorporating text in shiny applications.
To work with \code{usethis::use_data} the excel file should have named tabs,
}
\seealso{
\code{\link{prepare_text_from_excel}} converts texts provided via
  a single excel file into a structure that \code{select_text} can work with. Used by this function on each file provided.
  
 \code{\link{select_text}} extracts individual texts by tab,
 section and, optionally, top-level names.
}
