% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_text.R
\name{prepare_nested_text_lists_from_xlsx}
\alias{prepare_nested_text_lists_from_xlsx}
\title{Create a named nested list of dataframes containing text content, from excel files and a vector of names.}
\usage{
prepare_nested_text_lists_from_xlsx(paths, names)
}
\arguments{
\item{paths}{}

\item{names}{}
}
\description{
Uses \code{prepare_texts_from_excel} to make a nested list containing text 
content extractable by \emph{section}. Requires that \emph{paths} and 
\emph{names} are vectors of equal length. 
Output works with \code{select_text} where \emph{names} gives the 
\emph{top_level} parameter to \code{select_text}.  vector of names an excel file into a list of dataframes. Works with \code{usethis::use_data} and
 \code{select_text} to provide a simple means of incorporating text in shiny applications.
To work with \code{usethis::use_data} the excel file should have named tabs,
}
