% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_text.R
\name{prepare_text_from_excel}
\alias{prepare_text_from_excel}
\title{Generate a list of dataframes containing text content, from excel}
\usage{
prepare_text_from_excel(path_to_file)
}
\arguments{
\item{path_to_file}{path to the file containing text content}
}
\value{
a list of dataframes
}
\description{
Converts an excel file into a list of dataframes. Works with \code{usethis::use_data} and
 \code{select_text} to provide a simple means of incorporating text in shiny applications.
To work with \code{usethis::use_data} the excel file should have named tabs, 
each tab having 2 columns, \strong{section} and \strong{text}. \emph{text} is the content,
\emph{section} is used by \code{select_text} to extract the relevant section.
Texts provided in \emph{text} are wrapped in the HTML paragraph tag <p>.
}
\examples{
example_path <- system.file("extdata", "text_function_example.xlsx", package = "icesUtils")
prepare_text_from_excel(path_to_file = example_path)
}
\seealso{
\code{\link[icesUtils]{prepare_nested_text_lists_from_excel}} converts texts
  in multiple files into a hierarchical structure that \code{select_text} can work with.
  
 \code{\link[icesUtils]{select_text}} extracts individual texts by tab,
 section and, optionally, top-level names.
}
