% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_select_text.R
\name{select_text}
\alias{select_text}
\title{´Get text to display from list of dfs.}
\usage{
select_text(list_of_texts, tab, section, ecoregion = NULL)
}
\arguments{
\item{list_of_texts}{a list of dataframes each containing 'section' and 'text' columns}

\item{tab}{a character vector - the name of a list entry}

\item{section}{a character vector indicating which row to extract}

\item{ecoregion}{optional argument for top level sorting}
}
\value{
A character string
}
\description{
display_text´ subsets a list of dataframes and extracts the relevant section of text for tab and section
If ecoregion is supplied, accepts a named list of lists, and filters the top level list by ecoregion name.
}
\examples{
greetings <- data.frame(section = c("welcome", "goodbye"), 
text = c("Hello world", "Thanks for stopping by"))
content <- data.frame(section = "help", text = "Use the function as shown here")
texts <- list(greetings=greetings, content = content)
select_text(list_of_texts = texts, "greetings", "welcome")
select_text(list_of_texts = texts, "content", "help")
select_text(list_of_texts = texts, "greetings", "goodbye")

}
\author{
Neil Maginnis
}
